refactor: WORLD/CHANNEL에서 common의 redis/mysql 사용으로 변경 - redis/mysql 방식변경  1. Init 호출(config 인자로 삽입)  2. GetInstance 호출 (Init안될시 호출 안됨)
