+++ SERVER/include/COMMON/db/MySqlConnectionPool.h
@@ -1,20 +1,24 @@
 #pragma once
-#include <mysql/mysql.h>
-#include <string>
+#include "common.h"
 #include <queue>
+#include "ConfigLoader.h"
 
-//SingleTone
+#define MYSQL_POOL_SIZE 8
+
 class MySqlConnectionPool
 {
 public:
     static MySqlConnectionPool *GetInstance();
-
-    int Init(const int pool_size);
-    MYSQL*  GetConnection();
+    static int Init(const MySqlConfig& mysqlConfig, const int pool_size = MYSQL_POOL_SIZE);
+    
+    MYSQL* GetConnection();
     int ReleaseConnection(MYSQL*);
+
 private:
+    explicit MySqlConnectionPool(const MySqlConfig& mysqlConfig, const int pool_size);
+    int GetPoolSize() const;
     std::queue<MYSQL*> m_pool;
-    MySqlConnectionPool(const int pool_size);
     ~MySqlConnectionPool();
-    static MySqlConnectionPool *m_instance;
+
+    static MySqlConnectionPool* m_instance;
 };
