+++ SERVER/src/COMMON/db/MySqlConnectionPool.cpp
@@ -1,53 +1,82 @@
 #include "MySqlConnectionPool.h"
-#include "K_slog.h"
 
-#define POOL_SIZE 8
-#define HOST "127.0.0.1"
-#define USER "root"
-#define PASSWD "1234"
-#define DB "game"
-#define PORT 3306
+MySqlConnectionPool *MySqlConnectionPool::m_instance=nullptr;
 
-MySqlConnectionPool *MySqlConnectionPool::m_instance = nullptr;
+int MySqlConnectionPool::Init(const MySqlConfig& mysqlConfig, const int pool_size)
+{
+    if (m_instance != nullptr)
+    {
+        K_slog_trace(K_SLOG_ERROR, "[%s][%d] Already Init ", __FUNCTION__, __LINE__);
+        return -1;
+    }
 
-int MySqlConnectionPool::Init(const int pool_size)
+    MySqlConnectionPool* client = new MySqlConnectionPool(mysqlConfig, pool_size);
+    if (client == nullptr)
+    {
+        K_slog_trace(K_SLOG_ERROR, "[%s][%d] Memory error(new MySqlConnectionPool(mysqlConfig, pool_size)) ", __FUNCTION__, __LINE__);
+        return -1;
+    }
+
+    if (client->GetPoolSize() == 0)
+    {
+        K_slog_trace(K_SLOG_ERROR, "[%s][%d] connect fail host=%s, port=%d", __FUNCTION__, __LINE__, mysqlConfig.host.c_str(), mysqlConfig.port);
+        K_slog_trace(K_SLOG_ERROR, "[%s][%d] connect fail user=%s, database=%s", __FUNCTION__, __LINE__, mysqlConfig.user.c_str(), mysqlConfig.database.c_str());
+        delete client;
+        return -1;
+    }
+    
+    m_instance = client;
+    return 0;
+}
+
+int MySqlConnectionPool::GetPoolSize() const
+{
+    return m_pool.size();
+}
+
+MySqlConnectionPool::MySqlConnectionPool(const MySqlConfig& mysqlConfig, const int pool_size)
 {
     int cnt = 0;
-    //db커넥션 반복문으로 생성
+
     for (int i = 0; i < pool_size; i++)
     {
-        MYSQL* conn = nullptr;
-        conn = mysql_init(nullptr);
+        MYSQL* conn = mysql_init(nullptr);
         if (!conn)
         {
-            K_slog_trace(K_SLOG_ERROR, "[%s][%d] mysql_init", __FUNCTION__, __LINE__);
+            K_slog_trace(K_SLOG_ERROR, "[%s][%d] mysql_init ERROR", __FUNCTION__, __LINE__);
             continue;
         }
-        if (!mysql_real_connect(conn, HOST, USER, PASSWD, DB, PORT, nullptr, 0))
-        {
-             K_slog_trace(K_SLOG_ERROR,
-                "[%s][%d] mysql_real_connect failed: %s",
-                __FUNCTION__, __LINE__, mysql_error(conn));
 
+        MYSQL* result = mysql_real_connect(
+            conn,
+            mysqlConfig.host.c_str(),
+            mysqlConfig.user.c_str(),
+            mysqlConfig.password.c_str(),
+            mysqlConfig.database.c_str(),
+            mysqlConfig.port,
+            nullptr,
+            0
+        );
+
+        if (result == nullptr)
+        {
+            K_slog_trace(K_SLOG_ERROR, "[%s][%d] mysql_real_connect failed : %s",
+                         __FUNCTION__, __LINE__, mysql_error(conn));
             mysql_close(conn);
             continue;
         }
-        mysql_autocommit(conn, 0);
+
+
         m_pool.push(conn);
         cnt++;
     }
 
-    K_slog_trace(K_SLOG_DEBUG, "[%s][%d] db pool created[%d]", __FUNCTION__, __LINE__, cnt);
-    return 0;
+    K_slog_trace(K_SLOG_TRACE, "[%s][%d] db pool created[%d]", __FUNCTION__, __LINE__, cnt);
 }
 
-MySqlConnectionPool::MySqlConnectionPool(const int pool_size)
-{
-    Init(pool_size);
-}
 MySqlConnectionPool::~MySqlConnectionPool()
 {
-    if (m_instance)
+    if(m_instance)
     {
         delete m_instance;
         m_instance = nullptr;
@@ -56,9 +85,10 @@ MySqlConnectionPool::~MySqlConnectionPool()
 
 MySqlConnectionPool* MySqlConnectionPool::GetInstance()
 {
-    if (m_instance == nullptr)
+    if(m_instance == nullptr)
     {
-        m_instance = new MySqlConnectionPool(POOL_SIZE);
+        K_slog_trace(K_SLOG_ERROR, "[%s][%d] Not Initialized mysql (first call MySqlConnectionPool::Init) ", __FUNCTION__, __LINE__);
+        return nullptr;
     }
 
     return m_instance;
@@ -66,6 +96,11 @@ MySqlConnectionPool* MySqlConnectionPool::GetInstance()
 
 MYSQL* MySqlConnectionPool::GetConnection()
 {
+    // m_pool이 생성안됐을 때 대비해서 안전 코드 생성
+    if(m_pool.empty())
+    {
+        return nullptr;
+    }
     MYSQL* conn = m_pool.front();
     m_pool.pop();
     return conn;
@@ -73,7 +108,7 @@ MYSQL* MySqlConnectionPool::GetConnection()
 
 int MySqlConnectionPool::ReleaseConnection(MYSQL* conn)
 {
-    if (conn != nullptr)
+    if(conn !=nullptr)
     {
         m_pool.push(conn);
     }
